requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchmetrics>=1.0.0
Pillow>=9.0.0
pandas>=1.3.0
tqdm>=4.65.0
numpy>=1.21.0
scipy>=1.7.0
einops>=0.6.0
lightning>=2.0.0
wandb>=0.15.0
kornia>=0.7.0
torch-fidelity>=0.3.0
transformers>=4.10.0
psutil>=5.8.0 